fix(build.rs): fix MaybeUninit usage & uninitialized integers (UB)
i changed Matrix to have zero-initialized integers because it was used
in so many methods, and would make things more convoluted to write
correctly that way. i believe this is fine because Matrix is only
constructed once (in create_zero_operator), so cost of initializing
128 bytes to zero should be entirely negligible (especially for a
build script).

on that note, it may be simpler to just zero-initialize everything. i
can only really see benefits to tightly optimizing initialization if
it's /very/ hot code, but this is a build script, and we would be at
most zero-initializing 8KiB.

Signed-off-by: Nissa <and-nissa@protonmail.com>